Rewrite segment drawing. Big code change, no user changes
The previous segment drawing was not well structured, making it hard to
understand and hard to modify. There was duplication of lseg* and rseg*.
And it did not separate the stage of collecting the active segments,
then assembling those collected items together. Instead it assembled as
it went along. That made it hard to add post-processing features, like
grouping segments or changing the look of one segment based on the
result of another.

The new code collects the active segments in a set of arrays. Then it
can do all the formatting without needing to inspect any external stuff
once those arrays are done. This enables using the same exact code for
left-to-right and right-to-left prompts. It simplifies a lot of the
logic for picking the right colors and shapes for separators. And it
enables fancy backwards-facing separators and other flashiness later on.

A test tool, util/draw-segments-test, has the same basic logic as the
actual prompt, and may be useful for prototyping new features later on.