Fix prompt_newt_defaults bug with empty context
It's possible to set a default with no extra context; that is, it
applies to ':prompt-theme:newt:*' itself. The most common are the
'left' and 'right' styles, for example. There was a display bug in
prompt_newt_default causing it to treat the context list as a string
instead of an array, giving "':prompt-theme:newt:*:lef' t".