Don't clobber existing zstyle settings (:vcs_info:*)
This gets most of the way toward blending existing zstyle settings with
newt's defaults. Now the user's ':vcs_info:*' check-for-changes is
kept, rather than being clobbered unconditionally.

In addition, a snapshot of styles in use is saved to an environment
variable, so if the environment is inherited by another newt prompt, it
will use the same styles and behave the same. This is particularly nice
when using `sudo -Es` to work as root for a while.